% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_palette.R
\name{expand_palette}
\alias{expand_palette}
\title{Expand palette}
\usage{
expand_palette(palette, amount = 0.2)
}
\arguments{
\item{palette}{A color palette (vector of hex colors) to be expanded}

\item{amount}{Amount of brightening of colors in palette (0 to 1)}
}
\value{
An expanded color palette
}
\description{
Inputs a color palette (vector) and
expands the colors by repeating them
in gradually lighter shades
}
