% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/su_palette.R
\name{su_palette}
\alias{su_palette}
\title{#' SU palette}
\usage{
su_palette(name, n, type = c("discrete", "continuous"), expand = FALSE)
}
\arguments{
\item{name}{Name of palette, options: \code{su_basic}, \code{su_expanded},
\code{su_dark}, \code{su_olive}, \code{su_sky},
\code{su_fire}, \code{su_water}}

\item{n}{Number of colors needed}

\item{type}{Continuous or discrete}

\item{expand}{Expand palette length (default is `FALSE`)}
}
\description{
Color palettes recreating the visual identity of SU
(https://www.su.se/medarbetare/kommunikation/grafisk-manual)
}
