% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/su_palette.R
\docType{data}
\name{su_palettes}
\alias{su_palettes}
\title{SU palette}
\format{
An object of class \code{list} of length 7.
}
\usage{
su_palettes
}
\description{
Color palettes recreating the visual identity of SU
(https://www.su.se/medarbetare/kommunikation/grafisk-manual)
}
\details{
(code based on https://github.com/jrcalabrese/joycon)

Options: \code{su_basic}, \code{su_expanded},
\code{su_dark}, \code{su_olive}, \code{su_sky},
\code{su_fire}, \code{su_water}
}
\keyword{datasets}
