% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_su.R
\name{theme_su}
\alias{theme_su}
\title{SU ggtheme}
\usage{
theme_su(base_size = 14, base_family = "Verdana")
}
\arguments{
\item{base_size}{Default font size}

\item{base_family}{Default font family}
}
\description{
SU ggtheme
}
